requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asn1crypto==0.24.0        # via cryptography
certifi==2019.3.9         # via requests
cffi==1.12.2              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via flask
cryptography==2.6.1       # via pyjwt
flask==1.0.2
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
markupsafe==1.1.1         # via jinja2
nexmo==2.3.0
pycparser==2.19           # via cffi
pyjwt[crypto]==1.7.1      # via nexmo
python-dotenv==0.10.1
pytz==2018.9              # via nexmo
requests==2.21.0          # via nexmo
six==1.12.0               # via cryptography
tinydb==3.13.0
urllib3==1.24.1           # via requests
werkzeug==0.15.1          # via flask
